Module hotstuff_rs::block_sync::messages

Definitions for structured messages that are sent between replicas as part of the Block Sync protocol.

Messages

The Block Sync protocol defines two categories of messages:

Block Sync Protocol messages (BlockSyncMessage): exchanged between a sync client and sync server when the client is trying to sync with the server.

Block Sync Server Advertisements (BlockSyncAdvertiseMessage): periodically broadcasted by sync servers to update the sync clients on:

Their availability and commitment to providing blocks at least up to a given height (AdvertiseBlock).

Whether the quorum is making progress in a future view, as evidenced by the server’s local Highest PC (AdvertisePC).

Structs

AdvertiseBlock

A message periodically broadcasted by the sync server to:

AdvertisePC

A message periodically broadcasted by the sync server to let clients know about the Highest PhaseCertificate the server knows.

BlockSyncRequest

Sync request sent by a sync client to a sync server. The request includes:

BlockSyncResponse

Sync response sent by a sync server to a sync client requesting blocks. The response includes:

Enums

BlockSyncAdvertiseMessage

Messages periodically broadcasted by the sync server to update clients about their availability, commitment to providing blocks up to a given height, and knowledge of the latest consensus decision.

BlockSyncMessage

Messages exchanged between a sync server and a sync client when syncing.