Enum hotstuff_rs::block_sync::messages::BlockSyncAdvertiseMessage

pub enum BlockSyncAdvertiseMessage {
    AdvertiseBlock(AdvertiseBlock),
    AdvertisePC(AdvertisePC),
}

Messages periodically broadcasted by the sync server to update clients about their availability, commitment to providing blocks up to a given height, and knowledge of the latest consensus decision.

Variants

AdvertiseBlock(AdvertiseBlock)

AdvertisePC(AdvertisePC)

Implementations

impl BlockSyncAdvertiseMessage

pub(crate) fn advertise_block( me: &Keypair, chain_id: ChainID, highest_committed_block_height: BlockHeight ) -> Self

pub(crate) fn advertise_pc(highest_pc: PhaseCertificate) -> Self

pub fn chain_id(&self) -> ChainID

pub fn size(&self) -> u64