hotstuff_rs::block_tree::accessors
Module app

Special, read-only interface for querying the Block Tree used only by Apps.

Structs

AppBlockTreeView

View of the block tree made available to method calls on Apps.