hotstuff_rs::block_tree::accessors::internal
Enum BlockTreeError

pub enum BlockTreeError {
    KVGetError(KVGetError),
    KVSetError(KVSetError),
    BlockExpectedButNotFound {
        block: CryptoHash,
    },
}

Errors that may be encountered when reading or writing to the BlockTreeSingleton.

Variants

KVGetError(KVGetError)

Error when trying to get a value from the block tree’s underlying key value store.

KVSetError(KVSetError)

Error when trying set a value into block tree’s underlying key value store.

BlockExpectedButNotFound

Unable to find a block with the specific CryptoHash, even though an invariant that the block tree expects to be maintained suggests that the block should exist.

Fields

block: CryptoHash