hotstuff_rs::block_tree::invariants
Function extends_locked_pc_block

fn extends_locked_pc_block<K: KVStore>(
    pc: &PhaseCertificate,
    block_tree: &BlockTreeSingleton<K>,
) -> Result<bool, BlockTreeError>

Check whether pc belongs to the branch that extends from the block_tree’s locked_pc.block.

Procedure

In the phased mode, it is always the latest block that is locked, while in the pipelined mode, it is the grandparent of the newest block that is locked. Therefore, to see whether pc extends from locked_pc.block, we only need to check whether or not locked_pc.block is pc.block, or its parent, or its grandparent. We do not need to check deeper.