hotstuff_rs::block_tree::pluggables
Enum KVGetError

pub enum KVGetError {
    DeserializeValueError {
        key: Key,
        source: Error,
    },
    ValueExpectedButNotFound {
        key: Key,
    },
    Ed25519DalekError {
        key: Key,
        source: SignatureError,
    },
}

Error when trying to read a value corresponding to a given key from the key value store. The error may arise in the following circumstances:

The value corresponding to a given key cannot be deserialized into its expected type,

The value corresponding to a given key cannot be found,

There was an error when processing Ed25519Dalek keypairs or signatures.

Variants

DeserializeValueError

Fields

key: Key

source: Error

ValueExpectedButNotFound

Fields

key: Key

Ed25519DalekError

Fields

key: Key

source: SignatureError