hotstuff_rs::hotstuff::types
Struct PhaseCertificate

pub struct PhaseCertificate {
    pub chain_id: ChainID,
    pub view: ViewNumber,
    pub block: CryptoHash,
    pub phase: Phase,
    pub signatures: SignatureSet,
}

Cryptographic proof that at least a quorum of validators have voted for a Proposal or Nudge with the included chain_id, view, block, and phase.

Used to extend a block in the HotStuff subprotocol, and to optimistically advance to a new view in the pacemaker subprotocol.

Fields
chain_id: ChainIDview: ViewNumberblock: CryptoHashphase: Phasesignatures: SignatureSet
Implementations

impl PhaseCertificate

pub const fn genesis_pc() -> PhaseCertificate

Get the โGenesis Phase Certificateโ, the phase certificate that Blocks of height 0 (โGenesis Blocksโ) include as their block.justify.

The fields of this PhaseCertificate all take the โdefaultโ values of their respective types. For example, genesis_pc.chain_id is 0 regardless of the ChainID of the specific block tree, and genesis_pc.block is [0u8; 32].

pub fn is_genesis_pc(&self) -> bool

Check whether this PhaseCertificate is the genesis_pc.

pub fn is_block_justify(&self) -> bool

Check whether self.phase is a variant that is allowed for PhaseCertificates that serve as block.justify.

Returns true if phase.is_generic() or phase.is_decide(). Returns false otherwise.

pub fn is_nudge_justify(&self) -> bool

Check whether self.phase is a veriant that is allowed for PhaseCertificates that serve as nudge.justify.

Returns true if phase.is_prepare(), phase.is_precommit(), or phase.is_commit(). Returns false otherwise.