Trait hotstuff_rs::logging::Logger

pub(crate) trait Logger {
    // Required method
    fn get_logger() -> Box<dyn Fn(&Self) + Send>;
}

Implemented by event types. Used to get a closure that logs the event.

Required Methods

fn get_logger() -> Box<dyn Fn(&Self) + Send>

Returns a pointer to the default logging handler for a given event type.

Object Safety

This trait is not object safe.