hotstuff_rs
Module networking

Pluggable peer-to-peer (P2P) networking.

Modules

messages

Exhaustive enumerations around every message variant used in HotStuff-rs.

network

Trait for pluggable peer-to-peer (P2P) networking.

receiving 🔒

Functions and types for receiving messages from the P2P network.

sending 🔒

Functions and types for sending messages to the P2P network.