Module hotstuff_rs::networking::receiving

Structs

BlockSyncClientStub 🔒

A receiving end for sync responses. The BlockSyncClientStub::recv_response method returns the received response.

BlockSyncServerStub 🔒

A receiving end for sync requests. The BlockSyncServerStub::recv_request method returns the received request.

ProgressMessageBuffer 🔒

Message buffer intended for storing received ProgressMessages for future views. Its size is bounded by its capacity, and when the capacity is reached messages for highest views may be removed.

ProgressMessageStub 🔒

A receiving end for ProgressMessages.

Enums

BlockSyncRequestReceiveError

BlockSyncResponseReceiveError

ProgressMessageReceiveError 🔒

Traits

Cacheable 🔒

A cacheable message can be inserted into the progress message buffer.

Functions

start_polling 🔒

Spawn the poller thread, which polls the Network for messages and distributes them into receivers for: