hotstuff_rs::pacemaker
Module implementation

Event-driven implementation of the Pacemaker subprotocol.

Main type: Pacemaker.

Structs

Pacemaker 🔒

A single participant in the Pacemaker subprotocol.

PacemakerConfiguration 🔒

Configuration variables for the Pacemaker struct.

PacemakerState 🔒

In-memory state of a Pacemaker.

ViewInfo 🔒

Describes a view (most often the current view), in terms of its view number and its view deadline (the instant in time in which the view should end if no progress was made).

Enums

ExtendViewError

Enumerates the different ways a Pacemaker::extend_view call can fail.

PacemakerError

Enumerates the different ways a call to any of Pacemaker’s methods can fail.

UpdateViewError

Enumerates the different ways a Pacemaker::update_view call can fail.

Functions

epoch 🔒

Compute the current epoch based on the current view and the configured epoch_length.

is_epoch_change_view 🔒

Check whether view is an epoch-change view given the configured epoch_length.

select_leader

Deterministically select a replica in validator_set to become the leader of view using the Interleaved WRR algorithm.