hotstuff_rs::pacemaker
Module protocol

Implementation of the Pacemaker protocol, based on the Lewis-Pye View Synchronisation protocol and the Interleaved Weighted Round Robin leader selection mechanism.

The liveness of the HotStuff protocol is dependent on the Pacemaker module, which regulates how and when a replica advances its view, as well as determines which validator shall act as the leader of a given view.

View Synchronisation

The goal is to ensure that at any point all honest replicas should eventually end up in the same view and stay there for long enough to enable consensus through forming a PC. Just like the HotStuff SMR, the Pacemaker protocol is Byzantine Fault Tolerant: eventual succesful view synchronization is guaranteed in the presence of n = 3f + 1 validators where at most f validators are Byzantine.

The Lewis-Pye Pacemaker achieves view synchronisation by dividing the sequence of views into epochs. The mechanism for advancing the view depends on whether a view is advanced within the same epoch or involves epoch change:

All-to-all broadcast in every epoch-change view (i.e., last view of an epoch) upon which replicas enter the next epoch and set their timeout deadlines for all views in the next epoch,

Advancing to a next view within the same epoch either on timeout or optimistically on receiving a PC for their current view.

The latter ensures synchronisation when timeouts are set in a uniform manner and when leaders are honest, and the former serves as a fallback mechanism in case views fall out of sync.

This protocol deviates from Lewis-Pye in two fundamental ways:

Epoch length is configurable, rather than equal to f+1. This is to enable dynamic validator sets.

TimeoutVotes include a highest_tc the sender knows. This provides a fallback mechanism for helping validators lagging behind on epoch number catch up with the validators ahead.

Leader Selection

Leaders are selected according to Interleaved Weighted Round Robin algorithm. This ensures that:

The frequency with which a validator is selected as a leader is proportional to the validator’s power,

Validators are selected as leaders in an interleaved manner: unless a validator has more power than any other validator, it will never act as a leader for more than one consecutive view.

Structs

Pacemaker 🔒

A Pacemaker protocol for Byzantine View Synchronization inspired by the Lewis-Pye View Synchronization protocol. Its PacemakerState is the authoritative source of information regarding the current view and its leader, and Algorithm should regularly query the Pacemaker for this information (ViewInfo), and propagate the information to hotstuff.

PacemakerConfiguration 🔒

Immutable parameters that determine the behaviour of the Pacemaker and should never change after a replica starts.

PacemakerState 🔒

Internal state of the Pacemaker. Keeps track of the timeouts allocated to current and future views (if any), and the timeout votes collected for the current view.

ViewInfo 🔒

Describes a view (most often the current view), in terms of its view number and its view deadline (the instant in time in which the view should end if no progress was made).

Enums

ExtendViewError

Extending the view can fail in two ways:

PacemakerError

The pacemaker can fail in two fundamental ways:

UpdateViewError

Updating the view can fail in two ways:

Functions

epoch 🔒

Compute the current epoch based on the current view and the configured epoch_length.

is_epoch_change_view 🔒

Check whether view is an epoch-change view given the configured epoch_length.

select_leader

Implements the Interleaved Weighted Round Robin algorithm for selecting a view leader. For internal use by the Pacemaker and PacemakerState methods.