Module hotstuff_rs::pacemaker::types

Definitions of types specific to the Pacemaker protocol.

Structs

TimeoutCertificate

Proof that at least a quorum of validators have sent a TimeoutVote for the same view. Required for advancing to a new epoch as part of the pacemaker protocol.

TimeoutVoteCollector 🔒

Helps leaders incrementally form TimeoutCertificates by combining votes for the same chain_id and view by replicas in a given validator set.