hotstuff_rs::replica
Struct Replica

pub struct Replica<K: KVStore> {
    block_tree_camera: BlockTreeCamera<K>,
    poller: Option<JoinHandle<()>>,
    poller_shutdown: Sender<()>,
    algorithm: Option<JoinHandle<()>>,
    algorithm_shutdown: Sender<()>,
    block_sync_server: Option<JoinHandle<()>>,
    block_sync_server_shutdown: Sender<()>,
    event_bus: Option<JoinHandle<()>>,
    event_bus_shutdown: Option<Sender<()>>,
}

A handle to the background threads of a HotStuff-rs replica. When this value is dropped, all background threads are gracefully shut down.

Fields
block_tree_camera: BlockTreeCamera<K>poller: Option<JoinHandle<()>>poller_shutdown: Sender<()>algorithm: Option<JoinHandle<()>>algorithm_shutdown: Sender<()>block_sync_server: Option<JoinHandle<()>>block_sync_server_shutdown: Sender<()>event_bus: Option<JoinHandle<()>>event_bus_shutdown: Option<Sender<()>>
Implementations

impl<K: KVStore> Replica<K>

pub fn initialize( kv_store: K, initial_app_state: AppStateUpdates, initial_validator_set_state: ValidatorSetState, )

Initializes the replica’s block tree with the intial app state updates and validator set updates.

pub fn block_tree_camera(&self) -> &BlockTreeCamera<K>

Get a BlockTreeCamera.