Module hotstuff_rs::state

Persistent state of a replica.

Modules

app_block_tree_view

Special, read-only interface for querying the Block Tree used only by Apps.

block_tree

Internal read-and-write handle used by the algorithm thread to mutate the Block Tree.

block_tree_snapshot

General purpose, read-only interface for querying the Block Tree.

invariants

Rules and predicates that help with maintaining the invariant properties of the Block Tree data structure.

kv_store

Pluggable key-value storage.

paths

Byte-prefixes that specify where each Block Tree variable is stored in the user-provided key-value store.

write_batch

This module defines the WriteBatch trait for batching writes to the key-value store, and its implementation for the block tree, the BlockTreeWriteBatch.