Module hotstuff_rs::types

Types and traits that are used across multiple sub-protocols or components of HotStuff-rs.

Other types and traits, specific to single components of HotStuff-rs protocols, can be found in the “types” submodules of their components, e.g., crate::hotstuff::types.

Modules

block

Block type and its methods.

crypto_primitives

Cryptographic primitives.

data_types

Types that exist only to store bytes, and do not have any major “active” behavior.

signed_messages

Signed messages, votes, and certificates.

update_sets

validator_set

Types that store information about validator sets or updates to validator sets.