Struct hotstuff_rs::types::signed_messages::ActiveCollectorPair

pub(crate) struct ActiveCollectorPair<CL: Collector> {
    cvs_collector: CL,
    pvs_collector: Option<CL>,
}

Struct that combines Collectors for the validator set(s) that could be considered “active” at any given ValidatorSetState, wrapping interaction with them behind a single interface.

Active Validator Sets

At any specific ValidatorSetState and in the execution of the HotStuff and Pacemaker subprotocols, the “Active Validator Sets” are the validator set(s) that should participate in voting (by sending PhaseVotes and TimeoutVotes, respectively).

At most two validator sets could be active at any given validator_set_state:

The Committed Validator Set (CVS) is always an active validator set,

While the Previous Validator Set (PVS) is active if-and-only-if the latest validator set update is not decided (i.e., validator_set_state.update_decided()).

Because Votes do not indicate which validator set they are a part of, ActiveCollectorPair’s collect method tries to collect any vote it receives with both collectors the pair may currently contain in turn.

Usage

Use new to create a ActiveCollectorPair for a specific ChainID, View, and the current ValidatorSetState. Then, call collect on it any time a Vote arrives.

Call update_validator_sets whenever the current ValidatorSetState changes. When the current ViewNumber changes, discard the ActiveCollectorPair and create a new one using the current view.

Fields
cvs_collector: CL

Collector collecting votes from the current Committed Validator Set (CVS).

pvs_collector: Option<CL>

Collector collecting votes from the current Previous Validator Set (PVS). Is set to None if validator_set_state.update_decided().

Implementations

impl<CL: Collector> ActiveCollectorPair<CL>

pub(crate) fn new( chain_id: ChainID, view: ViewNumber, validator_set_state: &ValidatorSetState, ) -> Self

Create an ActiveCollectorPair for chain_id, view, and validator_set_state.

pub(crate) fn collect( &mut self, signer: &VerifyingKey, message: CL::Vote, ) -> Option<CL::Certificate>

Collect message with both collectors in this ActiveCollectorPair.

pub(crate) fn update_validator_sets( &mut self, latest_vss: &ValidatorSetState, ) -> bool

Update the ActiveCollectorPair’s active validator sets given the latest current validator_set_state.

If validator_set_state is different from the latest ValidatorSetState known by the ActiveCollectorPair, the collectors will be updated and this method will return true. Otherwise calling this method is a no-op and returns false.