Trait hotstuff_rs::types::signed_messages::Certificate

pub(crate) trait Certificate {
    type Vote: Vote;

    // Required methods
    fn is_correct<K: KVStore>(
        &self,
        block_tree: &BlockTree<K>
    ) -> Result<bool, BlockTreeError>;
    fn is_correctly_signed(&self, validator_set: &ValidatorSet) -> bool;

    // Provided method
    fn quorum(validator_set_power: TotalPower) -> TotalPower { ... }
}

Data types that aggregate multiple Votes of the same type into evidence that a quorum of validators in a particular validator set supports a particular decision.

Required Associated Types

type Vote: Vote

The specific Vote type that this Certificate aggregates into one value.

Required Methods

fn is_correct<K: KVStore>( &self, block_tree: &BlockTree<K> ) -> Result<bool, BlockTreeError>

Check whether the certificate is “correct” ( i.e., whether it can serve as evidence that a particular decision has been approved by the quorum of validators assigned to collectively make the decision), given the current block_tree.

Guidelines for implementation

Implementations of is_correct should generally execute the following three steps (in addition to any specialized steps needed to check the correctness of the specific implementing type):

Get ValidatorSetState from block_tree.

Decide whether the certificate should be tested against the Committed Validator Set, or the Previous Validator Set, or against both.

Call is_correctly_signed on the certificate, passing the committed validator set, the previous validator set, or both (in separate calls).

fn is_correctly_signed(&self, validator_set: &ValidatorSet) -> bool

Check whether the certificate is correctly signed by a quorum of validators in the given validator_set.

In general, this method should only be called from inside is_correct. Other code should call is_correct instead of calling this method.

Provided Methods

fn quorum(validator_set_power: TotalPower) -> TotalPower

Compute the minimum voting power that a certificate produced by a validator set with validator_set_power total power must contain in order for it to correspond to a “quorum” of validators.

“Minimum” here is understood in the inclusive sense, a certificate corresponds to a quorum if its power is greater than or equal to the return value of quorum.

Uniqueness of quorums

The quorum size of a validator set with the total power validator_set_power is ceil(validator_set_power * 2/3).

This exact threshold (2/3rds) of the total power of a validator set is the minimum needed to guarantee the invariant that in any view, at most one Certificate (of each concrete kind) can be formed, given that at most 1/3rds of the total power could be Byzantine.

To see why, consider the “worst-case” scenario where, in the current view:

Just under 1/3rds (“1/3rds - 1”) of the total power of the current validator set is Byzantine.

Just over 2/3rds (“2/3rds + 1”) of the total power is honest.

To attack the uniqueness invariant in this scenario (i.e., to try and form two conflicting certificates), all Byzantine validators would (double-)vote for conflicting certificates A and B with the same view number. Both certificates would now have the power “1/3rds - 1”. Then, since honest validators only vote for one certificate of each concrete kind per view, one of the following will be the case:

Only A accumulates enough votes to constitute a quorum.

Only B accumulates enough votes to constitute a quorum.

Neither A and B accumulate enough votes to constitute a quorum.

In particular, the case where both A and B both constitute a quorum is impossible, because there simply isn’t enough remaining votes in “2/3rds + 1” to form two quorums. Critically, note that if the quorum threshold were any lower (e.g., just 2/3rds), this case would be possible, and the invariant would no longer hold.

Object Safety

This trait is not object safe.