hotstuff_rs::types::signed_messages
Trait SignedMessage

pub(crate) trait SignedMessage: Clone {
    // Required methods
    fn message_bytes(&self) -> Vec<u8>;
    fn signature_bytes(&self) -> SignatureBytes;

    // Provided method
    fn is_correct(&self, verifying_key: &VerifyingKey) -> bool { ... }
}

Data types that contain: 1. A message, and 2. A digital signature over said message whose correctness can be verified against a VerifyingKey.

Required Methods

fn message_bytes(&self) -> Vec<u8>

Get the bytes that are passed as input into the signing function to form the signature of the SignedMessage.

fn signature_bytes(&self) -> SignatureBytes

Get the signature of the SignedMessage.

Provided Methods

fn is_correct(&self, verifying_key: &VerifyingKey) -> bool

Verify that signature_bytes is a signature created by verifying_key over message_bytes.

Object Safety

This trait is not object safe.