hotstuff_rs::types::signed_messages
Trait Vote

pub(crate) trait Vote: SignedMessage {
    // Required methods
    fn chain_id(&self) -> ChainID;
    fn view(&self) -> ViewNumber;
}

Data types that indicate that a validator supports a particular decision about a particular chain_id and view.

Required Methods

fn chain_id(&self) -> ChainID

Get the chain_id of the chain that the Vote is about.

fn view(&self) -> ViewNumber

Get the view that the Vote is about.

Object Safety

This trait is not object safe.