hotstuff_rs::types::update_sets
Type Alias ValidatorSetUpdatesBytes

pub type ValidatorSetUpdatesBytes = UpdateSet<VerifyingKeyBytes, Power>;

Intermediate representation of ValidatorSetUpdates for safe serialization and deserialization.

Rationale

See the related section about ValidatorSetBytes.

Aliased Type

struct ValidatorSetUpdatesBytes {
    inserts: HashMap<[u8; 32], Power>,
    deletes: HashSet<[u8; 32]>,
}

Fields
inserts: HashMap<[u8; 32], Power>

Insertion updates that will be committed when a Block is committed.

deletes: HashSet<[u8; 32]>

Deletion updates that will be committed when a Block is committed.