Function hotstuff_rs::block_sync::client::is_sync_server_address

fn is_sync_server_address<K: KVStore>(
    verifying_key: &VerifyingKey,
    block_tree: &BlockTreeSingleton<K>,
) -> Result<bool, BlockSyncClientError>

Returns whether a given verifying key is recognised as a valid sync server address.

A replica is allowed to act as a sync server if either:

It is a member of the current committed validator set, or

One of the current speculative blocks proposes to add the replica to the validator set.

Recognising only committed and candidate validators as potential sync servers is an effective, though rather conservative solution to the problem of sybil attacks.