hotstuff_rs::hotstuff
Module messages

Messages sent between replicas as part of the HotStuff subprotocol.

Structs

NewView

Message sent by a replica to the leaders of the next view on view timeout to make them aware of the replica’s highest_pc.

Nudge

Message broadcasted by a leader in view to “nudge” other validators to participate in the voting phase after justify.phase in order to make progress in committing a validator-set-updating block in the block tree identified by chain_id.

PhaseVote

Message sent by a validator to a leader of view + 1 to indicate that the validator agrees to a specific Proposal or Nudge.

Proposal

Message broadcasted by a leader in view to propose to other validators that the block tree identified by chain_id be extended by inserting block.

Enums

HotStuffMessage

Every kind of message sent between replicas as part of the HotStuff subprotocol.