Module hotstuff_rs::logging

Functions that log out events.

The logs defined in this module are printed if the user enabled them via replica’s config.

HotStuff-rs logs using the log crate. To get these messages printed onto a terminal or to a file, set up a logging implementation.

Log message format

Log messages are CSVs (Comma Separated Values) with at least two values. The first two values are always:

The name of the event in PascalCase (defined in this module as constants).

The time the event was emitted (as number of seconds since the Unix Epoch).

The rest of the values differ depending on the kind of event. For example, the following snippet is how a ReceiveProposal is printed:

ReceiveProposal, 1701329264, Id5u7f6, fNGCJyk, 0

In the snippet:

The third value is the first seven characters of the Base64 encoding of the public address of the origin of the proposal.

The fourth value is the first seven characters of the Base64 encoding of the hash of the proposed block.

The fifth value is the height of the proposed block.

Constants

ADVANCE_VIEW

COLLECT_PC

COLLECT_TC

COMMIT_BLOCK

END_SYNC

INSERT_BLOCK

NEW_VIEW

NUDGE

PHASE_VOTE

PROPOSE

PRUNE_BLOCK

RECEIVE_ADVANCE_VIEW

RECEIVE_NEW_VIEW

RECEIVE_NUDGE

RECEIVE_PHASE_VOTE

RECEIVE_PROPOSAL

RECEIVE_SYNC_REQUEST

RECEIVE_TIMEOUT_VOTE

SEND_SYNC_RESPONSE

START_SYNC

START_VIEW

TIMEOUT_VOTE

UPDATE_HIGHEST_PC

UPDATE_HIGHEST_TC

UPDATE_LOCKED_PC

UPDATE_VALIDATOR_SET

VIEW_TIMEOUT

Traits

Logger 🔒

Implemented by event types. Used to get a closure that logs the event.

Functions

first_seven_base64_chars 🔒

progress_certificate_info 🔒

secs_since_unix_epoch 🔒