hotstuff_rs::networking
Module receiving

Functions and types for receiving messages from the P2P network.

Structs

BlockSyncClientStub 🔒

A receiving end for sync responses. The BlockSyncClientStub::recv_response method returns the received response.

BlockSyncServerStub 🔒

A receiving end for sync requests. The BlockSyncServerStub::recv_request method returns the received request.

ProgressMessageBuffer 🔒

Message buffer intended for storing received ProgressMessages for future views.

ProgressMessageStub 🔒

A receiving end for ProgressMessages.

Enums

BlockSyncRequestReceiveError

BlockSyncResponseReceiveError

ProgressMessageReceiveError 🔒

Traits

Cacheable 🔒

A cacheable message can be inserted into the progress message buffer.

Functions

start_polling 🔒

Spawn the poller thread, which polls the Network for messages and distributes them into receiver handles.