Trait hotstuff_rs::networking::receiving::Cacheable

pub(crate) trait Cacheable {
    // Required methods
    fn view(&self) -> ViewNumber;
    fn size(&self) -> u64;
}

A cacheable message can be inserted into the progress message buffer.

For this, we require that:

The message is associated with a view,

The message size is statically known and depends on a particular enum variant.

Required Methods

fn view(&self) -> ViewNumber

fn size(&self) -> u64