hotstuff_rs::types
Module crypto_primitives

Cryptographic primitives.

The definitions and re-exports in this module provide two categories of cryptographic primitives:

Cryptographic Hashes: provided by the [sha2] crate.

Digital Signatures: provided by the [ed25519_dalek] crate.

Structs

Keypair 🔒

A facade around SigningKey that implements method for sign-ing messages as well as a getter for the public key associated with the signing key.

Signature

Ed25519 signature.

SigningKey

ed25519 signing key which can be used to produce signatures.

VerifyingKey

An ed25519 public key.

Traits

Digest

Convenience wrapper trait covering functionality of cryptographic hash functions with fixed output size.

Signer

Sign the provided message bytestring using Self (e.g. a cryptographic key or connection to an HSM), returning a digital signature.

Verifier

Verify the provided message bytestring using Self (e.g. a public key)

Type Aliases

CryptoHasher

SHA-256 hasher.

SignatureError

Errors which may occur while processing signatures and keypairs.