Module hotstuff_rs::types::data_types

Types that exist only to store bytes, and do not have any major “active” behavior.

Structs

BlockHeight

Height of a block in the block tree.

BufferSize

Size of a buffer (in bytes).

ChainID

Number that uniquely identifies a blockchain.

ChildrenList

List of children of a particular block.

CryptoHash

32-byte cryptographic hash.

Data

Arbitrary, indexable data provided by an App to HotStuff-rs to be stored in a Block.

DataLen

Number of Datum stored in a block’s Data.

Datum

Unit of Data that can be queried individually from the block tree.

EpochLength

Configurable number of views in a Pacemaker epoch.

Power

Weight of a specific validator’s votes in consensus decisions.

SignatureBytes

Ed25519 digital signature.

SignatureSet

An ordered list of SignatureBytes from the same ValidatorSet.

TotalPower

Sum of the Powers of all validators in a ValidatorSet.

ViewNumber

View number. Starts at 0 and increases by 1 every time the Pacemaker times out.