hotstuff_rs::types::signed_messages
Trait Collector

pub(crate) trait Collector: Clone {
    type Vote: Vote;
    type Certificate: Certificate<Vote = Self::Vote>;

    // Required methods
    fn new(
        chain_id: ChainID,
        view: ViewNumber,
        validator_set: ValidatorSet,
    ) -> Self;
    fn collect(
        &mut self,
        signer: &VerifyingKey,
        vote: Self::Vote,
    ) -> Option<Self::Certificate>;
    fn chain_id(&self) -> ChainID;
    fn view(&self) -> ViewNumber;
    fn validator_set(&self) -> &ValidatorSet;
}

Types that progressively combine Votes with the same chain_id, view, and validator_set to form Certificates.

Required Associated Types

type Vote: Vote

The specific SignedMessage type that this Collector takes in as input.

type Certificate: Certificate<Vote = Self::Vote>

The specific Certificate type that this Collector returns as output.

Required Methods

fn new(chain_id: ChainID, view: ViewNumber, validator_set: ValidatorSet) -> Self

Create a new instance of the Collector, configuring it to collect SignedMessages for the specified chain_id, and view, and signed by a member of validator_set.

fn collect( &mut self, signer: &VerifyingKey, vote: Self::Vote, ) -> Option<Self::Certificate>

Collect a vote signed by signer, returning a Certificate if a quorum of matching Votes from the Collector’s configured validator_set has been collected.

No-ops

Calling this method is a no-op if:

vote.chain_id != self.chain_id()

vote.view != self.view()

Preconditions

vote.is_correct(signer).

fn chain_id(&self) -> ChainID

Get the ChainID of the chain that this Collector is currently configured to collect Votes about.

fn view(&self) -> ViewNumber

Get the View that this Collector is currently configured to collect Votes about.

fn validator_set(&self) -> &ValidatorSet

Get the ValidatorSet that this Collector is currently configured to collect Votes from.

Object Safety

This trait is not object safe.