hotstuff_rs::types::update_sets
Type Alias ValidatorSetUpdates

pub type ValidatorSetUpdates = UpdateSet<VerifyingKey, Power>;

Set of updates to the validator that are applied when a block is committed.

Aliased Type

struct ValidatorSetUpdates {
    inserts: HashMap<VerifyingKey, Power>,
    deletes: HashSet<VerifyingKey>,
}

Fields
inserts: HashMap<VerifyingKey, Power>

Insertion updates that will be committed when a Block is committed.

deletes: HashSet<VerifyingKey>

Deletion updates that will be committed when a Block is committed.