Module hotstuff_rs::types::validator_set

Types that store information about validator sets or updates to validator sets.

Structs

Signature

Ed25519 signature.

SigningKey

ed25519 signing key which can be used to produce signatures.

ValidatorSet

Stores the identities of validators and their voting powers.

ValidatorSetBytes 🔒

Intermediate representation of ValidatorSet for safe serialization and deserialization.

ValidatorSetState

Collection of basic information about the current (i.e., committed) validator set, and the validator set history.

VerifyingKey

An ed25519 public key.

Enums

ValidatorSetUpdatesStatus

Wraps around ValidatorSetUpdates, providing additional information on whether the updates have already been applied or not.

ValidatorSetUpdatesStatusBytes

Intermediate representation of ValidatorSetUpdatesStatus for safe serialization and deserialization.